extract file explorer from editorpanel
