+++ client/src/app/components/EditorPanel.tsx
@@ -3,7 +3,7 @@ import { useEditorPanel } from './editorPanel/useEditorPanel';
 import { UIEditorPanel } from './editorPanel/UIEditorPanel';
 
 export function EditorPanel() {
-	const { fileExplorerVisible } = useEditorPanel();
+	useEditorPanel();
 
-	return <UIEditorPanel fileExplorerVisible={fileExplorerVisible} />;
+	return <UIEditorPanel />;
 }
