+++ client/src/app/components/fileExplorer/Note.tsx
