+++ client/src/app/components/fileExplorer/UIFileExplorer.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { useStore } from '../../../hooks';
+import { useStore } from '@hooks';
 import { FileSystemItem } from '@types';
 import { Directory } from './Directory';
 import { Note } from './Note';
-import { getFileName } from '../../../../shared/utils';
-import PlusIcon from '../../../assets/icons/plus.svg';
-import FolderOpenIcon from '../../../assets/icons/folder-open.svg';
+import { getFileName } from '@shared/utils';
+import PlusIcon from '@assets/icons/plus.svg';
+import FolderOpenIcon from '@assets/icons/folder-open.svg';
 
 interface Props {
 	items: FileSystemItem[];
