+++ client/src/app/components/fileExplorer/note/UINote.tsx
